# Make TNJ faster by increasing cache locality of function pointers.