--- a/src/backend/cpu/TNJ/BufferNode.hpp
+++ b/src/backend/cpu/TNJ/BufferNode.hpp
@@ -58,19 +58,26 @@ namespace TNJ
                            });
         }
 
-        void calc(int x, int y, int z, int w)
+        void calc(int x, int y, int z, int w, int lim)
         {
             dim_t l_off = 0;
             l_off += (w < (int)m_dims[3]) * w * m_strides[3];
             l_off += (z < (int)m_dims[2]) * z * m_strides[2];
             l_off += (y < (int)m_dims[1]) * y * m_strides[1];
-            l_off += (x < (int)m_dims[0]) * x;
-            this->m_val = m_ptr[l_off];
+            T *in_ptr = m_ptr + l_off;
+            T *out_ptr = this->m_val.data();
+            for(int i = 0; i < lim; i++) {
+                out_ptr[i] = in_ptr[((x + i) < m_dims[0]) ? (x + i) : 0];
+            }
         }
 
-        void calc(int idx)
+        void calc(int idx, int lim)
         {
-            this->m_val = m_ptr[idx];
+            T *in_ptr = m_ptr + idx;
+            T *out_ptr = this->m_val.data();
+            for(int i = 0; i < lim; i++) {
+                out_ptr[i] = in_ptr[i];
+            }
         }
 
         void getInfo(unsigned &len, unsigned &buf_count, unsigned &bytes)