--- a/src/backend/cpu/unary.hpp
+++ b/src/backend/cpu/unary.hpp
@@ -15,7 +15,12 @@
 
 namespace cpu
 {
-#define sign(in) std::signbit(in)
+
+template<typename T>
+T sign(T in)
+{
+    return T(std::signbit(in));
+}
 
 template<typename T>
 T sigmoid(T in)
@@ -23,56 +28,61 @@ T sigmoid(T in)
     return (1.0) / (1 + std::exp(-in));
 }
 
-#define UNARY_FN(op)                            \
-    template<typename T>                        \
-    struct UnOp<T, T, af_##op##_t>              \
-    {                                           \
-        T eval(T in)                            \
-        {                                       \
-            return op(in);                      \
-        }                                       \
-    };                                          \
+#define UNARY_OP_FN(op, fn)                         \
+    template<typename T>                            \
+    struct UnOp<T, T, af_##op##_t>                  \
+    {                                               \
+        void eval(TNJ::array<T> &out,               \
+                  const TNJ::array<T> &in, int lim) \
+        {                                           \
+            for (int i = 0; i < lim; i++) {         \
+                out[i] = fn(in[i]);                 \
+            }                                       \
+        }                                           \
+    };                                              \
 
-UNARY_FN(sin)
-UNARY_FN(cos)
-UNARY_FN(tan)
+#define UNARY_OP(op) UNARY_OP_FN(op, std::op)
 
-UNARY_FN(asin)
-UNARY_FN(acos)
-UNARY_FN(atan)
+UNARY_OP(sin)
+UNARY_OP(cos)
+UNARY_OP(tan)
 
-UNARY_FN(sinh)
-UNARY_FN(cosh)
-UNARY_FN(tanh)
+UNARY_OP(asin)
+UNARY_OP(acos)
+UNARY_OP(atan)
 
-UNARY_FN(asinh)
-UNARY_FN(acosh)
-UNARY_FN(atanh)
+UNARY_OP(sinh)
+UNARY_OP(cosh)
+UNARY_OP(tanh)
 
-UNARY_FN(round)
-UNARY_FN(trunc)
-UNARY_FN(sign )
-UNARY_FN(floor)
-UNARY_FN(ceil)
+UNARY_OP(asinh)
+UNARY_OP(acosh)
+UNARY_OP(atanh)
 
-UNARY_FN(exp)
-UNARY_FN(sigmoid)
-UNARY_FN(expm1)
-UNARY_FN(erf)
-UNARY_FN(erfc)
+UNARY_OP(round)
+UNARY_OP(trunc)
+UNARY_OP_FN(sign, sign)
+UNARY_OP(floor)
+UNARY_OP(ceil)
 
-UNARY_FN(log)
-UNARY_FN(log10)
-UNARY_FN(log1p)
-UNARY_FN(log2)
+UNARY_OP(exp)
+UNARY_OP_FN(sigmoid, sigmoid)
+UNARY_OP(expm1)
+UNARY_OP(erf)
+UNARY_OP(erfc)
 
-UNARY_FN(sqrt)
-UNARY_FN(cbrt)
+UNARY_OP(log)
+UNARY_OP(log10)
+UNARY_OP(log1p)
+UNARY_OP(log2)
 
-UNARY_FN(tgamma)
-UNARY_FN(lgamma)
+UNARY_OP(sqrt)
+UNARY_OP(cbrt)
 
-#undef UNARY_FN
+UNARY_OP(tgamma)
+UNARY_OP(lgamma)
+
+#undef UNARY_OP
 #undef sign
 
     template<typename T, af_op_t op>
@@ -87,15 +97,18 @@ UNARY_FN(lgamma)
 
 #define iszero(a) ((a) == 0)
 
-#define CHECK_FN(name ,op)                      \
-    template<typename T>                        \
-    struct UnOp<char, T, af_##name##_t>         \
-    {                                           \
-        char eval(T in)                         \
-        {                                       \
-            return op(in);                      \
-        }                                       \
-    };                                          \
+#define CHECK_FN(name ,op)                          \
+    template<typename T>                            \
+    struct UnOp<char, T, af_##name##_t>             \
+    {                                               \
+        void eval(TNJ::array<char> &out,            \
+                  const TNJ::array<T> &in, int lim) \
+        {                                           \
+            for (int i = 0; i < lim; i++) {         \
+                out[i] = op(in[i]);                 \
+            }                                       \
+        }                                           \
+    };                                              \
 
     CHECK_FN(isinf, std::isinf)
     CHECK_FN(isnan, std::isnan)