- [x] Fix failures in non linear "jit"
- [x] Investigate other vector lengths.
- [x] Move calc for loop into eval
- [x] Align m_val vector
--------------------------
- Fixes https://github.com/arrayfire/arrayfire/issues/1226