@pavanky 2017-07-31
```
 $ ./examples/financial/black_scholes_options_cpu 
ArrayFire v3.5.1 (CPU, 64-bit Linux, build d71b21e3)
[0] Intel: Intel(R) Core(TM) i5-6500 CPU @ 3.20GHz, 15985 MB, Max threads(4) 
** ArrayFire Black-Scholes Example **
**          by AccelerEyes         **

Warming up done
Input Data Size =   200000. Mean GPU Time: 17.315600 ms
Input Data Size =   400000. Mean GPU Time: 33.297000 ms
Input Data Size =   600000. Mean GPU Time: 49.700000 ms
Input Data Size =   800000. Mean GPU Time: 65.396600 ms
Input Data Size =  1000000. Mean GPU Time: 82.264400 ms
Input Data Size =  1200000. Mean GPU Time: 98.945900 ms
Input Data Size =  1400000. Mean GPU Time: 114.548300 ms
Input Data Size =  1600000. Mean GPU Time: 130.762700 ms
Input Data Size =  1800000. Mean GPU Time: 147.738500 ms
Input Data Size =  2000000. Mean GPU Time: 163.576800 ms
 $ LD_PRELOAD=/home/pavan/Workspace/local/lib/libafcpu.so.3.5.0 ./examples/financial/black_scholes_options_cpu 
ArrayFire v3.5.0 (CPU, 64-bit Linux, build a0dd4f92)
[0] Intel: Intel(R) Core(TM) i5-6500 CPU @ 3.20GHz, 15985 MB, Max threads(4) 
** ArrayFire Black-Scholes Example **
**          by AccelerEyes         **

Warming up done
Input Data Size =   200000. Mean GPU Time: 85.783400 ms
Input Data Size =   400000. Mean GPU Time: 170.254600 ms
Input Data Size =   600000. Mean GPU Time: 252.257600 ms
Input Data Size =   800000. Mean GPU Time: 335.771700 ms
Input Data Size =  1000000. Mean GPU Time: 419.037800 ms
Input Data Size =  1200000. Mean GPU Time: 504.373300 ms
Input Data Size =  1400000. Mean GPU Time: 590.496600 ms
Input Data Size =  1600000. Mean GPU Time: 674.670400 ms
Input Data Size =  1800000. Mean GPU Time: 754.990200 ms
Input Data Size =  2000000. Mean GPU Time: 839.069300 ms
```
black scholes is about 5x faster.