@pavanky 2017-07-31
Before
```
ArrayFire v3.5.0 (CPU, 64-bit Linux, build a0dd4f92)
[0] Intel: Intel(R) Core(TM) i5-6500 CPU @ 3.20GHz, 15985 MB, Max threads(4)
device:  0.68851 seconds to estimate  pi = 3.14180
  host:  0.61750 seconds to estimate  pi = 3.14126
```

After
```
ArrayFire v3.5.1 (CPU, 64-bit Linux, build d71b21e3)
[0] Intel: Intel(R) Core(TM) i5-6500 CPU @ 3.20GHz, 15985 MB, Max threads(4)
device:  0.47495 seconds to estimate  pi = 3.14180
  host:  0.61505 seconds to estimate  pi = 3.14126
```

Pi example is 1.44x faster.