@pavanky 2017-07-31
Looks like 256 may be optimum for the vector.