@pavanky 2017-08-01
After moving the loop into the eval:

pi_cpu
```
 $ ./examples/benchmarks/pi_cpu
ArrayFire v3.5.1 (CPU, 64-bit Linux, build 39aea578)
[0] Intel: Intel(R) Core(TM) i5-6500 CPU @ 3.20GHz, 15985 MB, Max threads(4)
device:  0.47724 seconds to estimate  pi = 3.14180
  host:  0.61517 seconds to estimate  pi = 3.14126
```

black_scholes_options_cpu
```
$ ./examples/financial/black_scholes_options_cpu 
ArrayFire v3.5.1 (CPU, 64-bit Linux, build 39aea578)
[0] Intel: Intel(R) Core(TM) i5-6500 CPU @ 3.20GHz, 15985 MB, Max threads(4) 
** ArrayFire Black-Scholes Example **
**          by AccelerEyes         **

Warming up done
Input Data Size =   200000. Mean GPU Time: 14.467300 ms
Input Data Size =   400000. Mean GPU Time: 28.436100 ms
Input Data Size =   600000. Mean GPU Time: 42.178500 ms
Input Data Size =   800000. Mean GPU Time: 55.509000 ms
Input Data Size =  1000000. Mean GPU Time: 69.781600 ms
Input Data Size =  1200000. Mean GPU Time: 82.447800 ms
Input Data Size =  1400000. Mean GPU Time: 96.588500 ms
Input Data Size =  1600000. Mean GPU Time: 110.566600 ms
Input Data Size =  1800000. Mean GPU Time: 123.624100 ms
Input Data Size =  2000000. Mean GPU Time: 137.566300 ms
```

black_scholes is about 20% since last update and 6x faster overall.