@pavanky 2017-08-01
@umar456 this is ready. Specializing for `VECTOR_SIZE == lim` is not improving the performance.