@pavanky 2017-08-01
build arrayfire ci